% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easystafe-package.R
\docType{package}
\name{easystafe-package}
\alias{easystafe}
\alias{easystafe-package}
\title{easystafe: Processamento de Extractos do e-SISTAFE}
\description{
Ferramentas para processar extractos do e-SISTAFE, incluindo deduplicação hierárquica e extração automática de metadados.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://moz-gpe.github.io/easystafe/}
}

}
\author{
\strong{Maintainer}: Joseph Lara \email{joseph.lara@giz.de}

}
\keyword{internal}
