% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easystafe-package.R
\docType{package}
\name{easystafe}
\alias{easystafe-package}
\alias{easystafe}
\title{easystafe: Ferramentas para Processamento de Extractos do e-SISTAFE}
\description{
O pacote \strong{easystafe} fornece funções para processamento, deduplicação,
extração de metadados e harmonização de extractos orçamentais provenientes
do sistema e-SISTAFE. Inclui um pipeline robusto para preparação de dados
para análise financeira, monitoria de execução e integração com workflows
baseados em R.
}
\details{
Este pacote encontra-se atualmente em fase \strong{experimental}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://moz-gpe.github.io/easystafe/}
}

}
\author{
\strong{Maintainer}: Joe Lara \email{joseph.lara@giz.de}

Authors:
\itemize{
  \item Damiano Stella \email{damiano.stella@giz.de}
}

}
\keyword{internal}
